base.indxg2p {pbdBASE}	R Documentation

indxg2p

Description

Computes the process coordinate which contains the entry of a distributed matrix specified by a global index INDXGLOB. Simplified reimplementation of the ScaLAPACK aux INDXG2P function.

Usage

base.indxg2p(INDXGLOB, NB, NPROCS)

Arguments

INDXGLOB
Global index.

NB
Block size.

NPROCS
Total number of processors over which matrix is distributed.

Details

For advanced users only.

[Package pbdBASE version 0.5-0 Index]