base.minctxt {pbdBASE}	R Documentation

Get BLACS Context Grid Information

Description

Finds the smallest integers for creating a new BLACS context.

Usage

base.minctxt(after = 0)

Arguments

after
ignores all values below this integer as possibilities

Details

For advanced users only.

Returns the smallest integer which could become a new BLACS context value.

For example, if contexts 0, 1, and 2 are taken, and after=0, then the function returns 3. If 0, 1, 2, and 5 are taken, the function returns 3 if after=0, but returns 6 if after=4.

The function is useful when a transitory grid is needed, such as for reading in data onto a subset of processors before distributing out to the full grid.

Value

Returns the minimum value.

[Package pbdBASE version 0.5-0 Index]