base.ownany {pbdBASE}	R Documentation

Determining Local Ownership of a Distributed Matrix

Description

For advanced users only.

Usage

base.ownany(dim, bldim, ICTXT = 0)

Arguments

dim
global dimension

bldim
blocking dimension

ICTXT
BLACS context

Details

A simple wrapper of numroc. The return is the answer to the question 'do I own any of the global matrix?'. Passing a distributed matrix is allowed, but often it is convenient to determine that information without even having a distributed matrix on hand. In this case, explicitly passing the appropriate information to the arguments dim=, bldim= (and CTXT= as necessary, since it defaults to 0) while leaving x missing will produce the desired result. See the examples below for more clarity.

The return for each function is local.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdBASE, quiet = TRUE)
init.grid()

iown <- ownany(dim=c(4, 4), bldim=c(2, 2), CTXT=0)
comm.print(iown, all.rank=T)

finalize()

## End(Not run)

[Package pbdBASE version 0.5-0 Index]