base.pdclvar {pbdBASE}	R Documentation

Column Variances

Description

Computes the variances of a ScaLAPCK-like distributed matrix. Significantly faster than using apply(), even in compared to the performance differences you would find comparing these two approaches using just base R.

Usage

base.pdclvar(x, descx)

Arguments

x
The matrix.

descx
ScaLAPACK descriptor array.

[Package pbdBASE version 0.5-0 Index]