base.rpdsyevx {pbdBASE}	R Documentation

rpdsyevx

Description

Genearlized eigenvalue problem.

Usage

base.rpdsyevx(jobz, range, n, a, desca, vl, vu, il, iu, abstol = 1e-08,
  orfac = 0.001)

Arguments

jobz
Control for if vectors/values/both are returned.

range
Parameter to determine the search criteria for eigenvalues.

n
Problem size.

a
Matrix.

desca
ScaLAPACK descriptor array.

vl, vu
Endpoints of the interval subset of the real line in which to search for eigenvalues, if specified by range.

il, iu
Eigenvalues with indices il, ..., iu will be found, if specified by range.

abstol
Absolute error tolerance for the eigenvalues.

orfac
Eigenvectors with eigenvalues below orfac*norm(a) of each other are reorthogonalized.

Details

For advanced users only.

[Package pbdBASE version 0.5-0 Index]