base.valid_context {pbdBASE}	R Documentation

BLACS Context Validation

Description

Checks if a supplied ICTXT is valid.

Usage

base.valid_context(ICTXT, ..., override = FALSE)

Arguments

ICTXT
BLACS context number.

...
Not used.

override
If override=FALSE, the context number will produce an error if it is any of the reserved contexts (0, 1, or 2).

[Package pbdBASE version 0.5-0 Index]