coords {pbdBASE}	R Documentation

Local to Global/Global to Local Indexing

Description

Get the local index given global information.

Usage

indxg2l(INDXGLOB, NB, IPROC, ISRCPROC, NPROCS)

indxl2g(INDXLOC, NB, IPROC, ISRCPROC, NPROCS)

Arguments

INDXGLOB
Global index.

NB
Block size.

IPROC
Coordinate of the process whose local info is to be determined.

ISRCPROC
The coordinate of the process that possesses the first row/column of the distributed matrix. That's always 0 pbdDMAT.

NPROCS
Total number of processors over which matrix is distributed.

INDXLOC
Local index.

Details

For advanced users only.

[Package pbdBASE version 0.5-0 Index]