gridexit {pbdBASE}	R Documentation

gridexit

Description

Frees a BLACS context.

Usage

base.gridexit(ICTXT, override = FALSE)

gridexit(ICTXT, override = FALSE)

Arguments

ICTXT
BLACS context number.

override
logical; if TRUE, ignores normal check preventing the closing of ICTXT values of 0, 1, and 2.

Details

For advanced users only.

The function frees the requested BLACS context. It is a trivial wrapper for the BLACS routine BLACS_GRIDEXIT. Also removes the object .__blacs_gridinfo_ICTXT.

Contexts 0, 1, and 2 can not be freed in this way unless the argument override=FALSE. This will probably break something and I do not recommend it.

Value

Silently returns 0 when successful. Silently returns 1 when requested ICTXT does not exist.

[Package pbdBASE version 0.5-0 Index]