pbdBASE-package {pbdBASE}	R Documentation

ScaLAPACK Wrappers and Utilities

Description

A package contains the basic methods for dealing with distributed data types, as well as the data types themselves.

Details

Package: 	 pbdBASE

Type: 	 Package

License: 	 MPL

LazyLoad: 	 yes

This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).

Author(s)

Drew Schmidt schmidt AT math.utk.edu, Wei-Chen Chen, George Ostrouchov, and Pragneshkumar Patel.

References

Programming with Big Data in R Website: http://r-pbd.org/

[Package pbdBASE version 0.5-0 Index]